Drop cache only for master branch

+++ index.js
@@ -26,6 +26,10 @@ handler.on('push', function (event) {
   console.log('Received a push event for %s to %s',
     event.payload.repository.name,
     event.payload.ref)
+  if (event.payload.ref !== 'refs/heads/master') {
+    return
+  }
+  console.log('master was deployed, purging cache...')
   request.del('https://api.cloudflare.com/client/v4/zones/' + CLOUDFLARE_ZONE + '/purge_cache')
     .set('X-Auth-Email', CLOUDFLARE_EMAIL)
     .set('X-Auth-Key', CLOUDFLARE_KEY)
